requirements.txt
attrs==21.2.0
backcall==0.2.0
blis==0.7.5
catalogue==2.0.6
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
coverage==6.0.2
cymem==2.0.6
debugpy==1.5.1
decorator==5.1.0
entrypoints==0.3
gensim==4.1.2
idna==3.3
iniconfig==1.1.1
ipykernel==6.4.2
ipython==7.28.0
ipython-genutils==0.2.0
jedi==0.18.0
Jinja2==3.0.2
bs4
lxml
jupyter-client==7.0.6
jupyter-core==4.8.1
language-check==1.1
language-tool-python==2.6.1
lxml==4.6.3
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
murmurhash==1.0.6
nest-asyncio==1.5.1
numpy==1.21.3
packaging==21.0
parso==0.8.2
pathy==0.6.0
pexpect==4.8.0
pickleshare==0.7.5
pluggy==1.0.0
preshed==3.0.6
prompt-toolkit==3.0.21
psycopg2-binary==2.9.1
ptyprocess==0.7.0
pubmed-parser==0.2.2
py==1.10.0
pycontractions==2.0.1
pydantic==1.8.2
pyemd==0.5.1
Pygments==2.10.0
pyparsing==3.0.0
pytest==6.2.5
pytest-cov==3.0.0
python-dateutil==2.8.2
pyzmq==22.3.0
requests==2.26.0
scipy==1.7.1
six==1.16.0
smart-open==5.2.1
spacy==3.1.3
spacy-legacy==3.0.8
srsly==2.4.2
thinc==8.0.11
toml==0.10.2
tomli==1.2.1
tornado==6.1
tqdm==4.62.3
traitlets==5.1.0
typer==0.4.0
typing-extensions==3.10.0.2
Unidecode==1.3.2
urllib3==1.26.7
wasabi==0.8.2
wcwidth==0.2.5
word2number==1.1
lxml
bs4
biopython
elsapy